refactor: rename to local_cleanup_queue
