+++ .vscode/settings.json
@@ -122,7 +122,8 @@
 		"scheduler_options.h": "c",
 		"route_config_parse.h": "c",
 		"route.h": "c",
-		"pool.h": "c"
+		"pool.h": "c",
+		"local_cleanup_queue.h": "c"
 	},
 	"files.exclude": {
 		"**/.git": true,
